+++ src/error.rs
@@ -30,7 +30,7 @@ pub enum FsDkrError {
     },
 
     #[error("Paillier verification proof failed for party {party_index:?}")]
-    PaillierVerificationError { party_index: usize },
+    PaillierVerificationError { party_index: u16 },
 
     #[error("A new party did not receive a valid index.")]
     NewPartyUnassignedIndexError,
@@ -39,5 +39,5 @@ pub enum FsDkrError {
     BroadcastedPublicKeyError,
 
     #[error("DLog proof failed for party {party_index:?}")]
-    DLogProofValidation { party_index: usize },
+    DLogProofValidation { party_index: u16 },
 }
